Check if an alias already exists
